Make the new functionality of the top level EndDash class tested all in one spec and move it down one level to the normal test directory
